Remove mockEmployeeData from Overview page so it can export as a React component
