+++ src/app/employee/mockDataOverview.ts
@@ -1,0 +1,57 @@
+export const mockEmployees = [
+    {
+      id: 1,
+      first_name: 'John',
+      last_name: 'Doe',
+      email: 'john.doe@example.com',
+      phone_number: '123-456-7890',
+      position: 'Software Engineer',
+      active: true,
+    },
+    {
+      id: 2,
+      first_name: 'Jane',
+      last_name: 'Smith',
+      email: 'jane.smith@example.com',
+      phone_number: '987-654-3210',
+      position: 'Product Manager',
+      active: false,
+    },
+    {
+      id: 3,
+      first_name: 'John',
+      last_name: 'Doe',
+      email: 'john.doe@example.com',
+      phone_number: '123-456-7890',
+      position: 'Software Engineer',
+      active: true,
+    },
+    {
+      id: 4,
+      first_name: 'Jane',
+      last_name: 'Smith',
+      email: 'jane.smith@example.com',
+      phone_number: '987-654-3210',
+      position: 'Product Manager',
+      active: false,
+    },
+    {
+      id: 5,
+      first_name: 'John',
+      last_name: 'Doe',
+      email: 'john.doe@example.com',
+      phone_number: '123-456-7890',
+      position: 'Software Engineer',
+      active: true,
+    },
+    {
+      id: 6,
+      first_name: 'Jane',
+      last_name: 'Smith',
+      email: 'jane.smith@example.com',
+      phone_number: '987-654-3210',
+      position: 'Product Manager',
+      active: false,
+    },
+  ];
+
